+++ R/plot_trajectory.R
@@ -10,7 +10,8 @@
 #' @param p centiles at which to draw the WHO polygons
 #' @param \ldots additional parameters passed to \code{\link{figure}}
 #' @examples
-#' fit <- fit_trajectory(subset(cpp, subjid == 2), y_var = "wtkg", method = "rlm")
+#' mod <- get_fit(cpp, y_var = "wtkg", method = "rlm")
+#' fit <- fit_trajectory(subset(cpp, subjid == 2), mod)
 #' plot(fit)
 #' plot(fit, center = TRUE)
 #' plot(fit, hover = c("wtkg", "bmi", "waz", "haz"))
@@ -44,22 +45,31 @@ plot.fittedTrajectory <- function(x, center = FALSE, x_range = NULL,
   ylab <- hbgd::hbgd_labels[[x$y_var]]
 
   if(center) {
-    for(el in c("xy", "fitgrid", "checkpoint"))
+    for(el in c("xy", "fitgrid", "checkpoint", "holdout")) {
       if(!is.null(x[[el]]))
         x[[el]]$y <- x[[el]]$y - who_centile2value(x[[el]]$x, p = 50,
           x_var = x$x_var, y_var = x$y_var, sex = x$sex)
+        if(!is.null(x[[el]]$yfit))
+          x[[el]]$yfit <- x[[el]]$yfit - who_centile2value(x[[el]]$x, p = 50,
+            x_var = x$x_var, y_var = x$y_var, sex = x$sex)
+    }
 
     ylab <- paste(ylab, "(WHO median-centered)")
   }
 
   fig <- figure(width = width, height = height,
-    xlab = hbgd::hbgd_labels[[x$x_var]], ylab = ylab, ...) %>%
+    xlab = hbgd::hbgd_labels[[x$x_var]], ylab = ylab, logo = NULL, ...) %>%
     ly_who(x = seq(x_range[1], x_range[2], length = 100), center = center,
       x_var = x$x_var, y_var = x$y_var, sex = x$sex, p = p) %>%
     rbokeh::ly_points(x, y, hover = hover, data = x$xy, color = "black")
-  if(!is.null(x$fitgrid))
+  if(!is.null(x$fitgrid)) {
     fig <- fig %>%
-      rbokeh::ly_lines(x, y, data = x$fitgrid, color = "black")
+      rbokeh::ly_lines(x, y, data = x$fitgrid, color = "black") %>%
+      rbokeh::ly_points(x, yfit, data = x$xy, color = "black", glyph = 19, size = 4)
+  }
+  if(!is.null(x$holdout))
+    fig <- fig %>%
+      rbokeh::ly_points(x, y, data = x$holdout, color = "red")
 
   if(!all(is.na(x$checkpoint$y)) && checkpoints) {
     x$checkpoint <- subset(x$checkpoint, !is.na(y))
@@ -86,10 +96,11 @@ plot.fittedTrajectory <- function(x, center = FALSE, x_range = NULL,
 #' @param z z-scores at which to draw the z-score bands
 #' @param \ldots additional parameters passed to \code{\link{figure}}
 #' @examples
-#' fit <- fit_trajectory(subset(cpp, subjid == 2), y_var = "wtkg", method = "rlm")
+#' mod <- get_fit(cpp, y_var = "wtkg", method = "rlm")
+#' fit <- fit_trajectory(subset(cpp, subjid == 2), mod)
 #' plot_z(fit)
 #' @export
-plot_z <- function(x, x_range = NULL, nadir = TRUE, width = 500, height = 520,
+plot_z <- function(x, x_range = NULL, nadir = FALSE, width = 500, height = 520,
   hover = NULL, checkpoints = TRUE, z = -3:0, ...) {
   if(is.null(x$xy$z))
     return(empty_plot("No z transformation data for this subject"))
@@ -112,13 +123,19 @@ plot_z <- function(x, x_range = NULL, nadir = TRUE, width = 500, height = 520,
   xlab <- hbgd::hbgd_labels[[x$x_var]]
   ylab <- paste(hbgd::hbgd_labels[[x$y_var]], "z-score")
 
-  fig <- figure(width = width, height = height, xlab = xlab, ylab = ylab, ...) %>%
+  fig <- figure(width = width, height = height,
+    xlab = xlab, ylab = ylab, logo = NULL, ...) %>%
     ly_zband(x = c(x_range[1], x_range[2]), z = z,
       color = ifelse(x$sex == "Male", "blue", "red")) %>%
     rbokeh::ly_points(x, z, hover = hover, data = x$xy, color = "black")
-  if(!is.null(x$fitgrid))
+  if(!is.null(x$fitgrid)) {
     fig <- fig %>%
-      rbokeh::ly_lines(x, z, data = x$fitgrid, color = "black")
+      rbokeh::ly_lines(x, z, data = x$fitgrid, color = "black") %>%
+      rbokeh::ly_points(x, zfit, data = x$xy, color = "black", glyph = 19, size = 4)
+  }
+  if(!is.null(x$holdout))
+    fig <- fig %>%
+      rbokeh::ly_points(x, z, data = x$holdout, color = "red")
 
   if(!all(is.na(x$checkpoint$y)) && checkpoints) {
     x$checkpoint <- subset(x$checkpoint, !is.na(y))
@@ -145,7 +162,8 @@ plot_z <- function(x, x_range = NULL, nadir = TRUE, width = 500, height = 520,
 #' @param height height of the plot
 #' @param \ldots additional parameters passed to \code{\link{figure}}
 #' @examples
-#' fit <- fit_trajectory(subset(cpp, subjid == 2), y_var = "wtkg", method = "rlm")
+#' mod <- get_fit(cpp, y_var = "wtkg", method = "rlm")
+#' fit <- fit_trajectory(subset(cpp, subjid == 2), mod)
 #' plot_velocity(fit)
 #' @export
 plot_velocity <- function(x, width = 500, height = 520, ...) {
@@ -155,8 +173,17 @@ plot_velocity <- function(x, width = 500, height = 520, ...) {
   xlab <- hbgd::hbgd_labels[[x$x_var]]
   ylab <- paste(hbgd::hbgd_labels[[x$y_var]], "growth velocity")
 
-  figure(width = width, height = height, xlab = xlab, ylab = ylab, ...) %>%
-    ly_lines(x, dy, data = x$fitgrid, color = "black")
+  # remove blip in velocity
+  xx <- x$fitgrid$x
+  dyy <- x$fitgrid$dy
+  ind <- which.min(abs(xx - 365.25 * 2))
+  if(abs(365.25 * 2 - xx[ind]) < 2 * diff(xx[1:2])) {
+    dyy[max(1, ind - 2):min(length(xx), ind + 2)] <- NA
+  }
+
+  figure(width = width, height = height,
+    xlab = xlab, ylab = ylab, logo = NULL, ...) %>%
+    ly_lines(xx, dyy, color = "black")
 }
 
 #' Plot a fitted trajectory's z-score velocity
@@ -166,7 +193,8 @@ plot_velocity <- function(x, width = 500, height = 520, ...) {
 #' @param height height of the plot
 #' @param \ldots additional parameters passed to \code{\link{figure}}
 #' @examples
-#' fit <- fit_trajectory(subset(cpp, subjid == 2), y_var = "wtkg", method = "rlm")
+#' mod <- get_fit(cpp, y_var = "wtkg", method = "rlm")
+#' fit <- fit_trajectory(subset(cpp, subjid == 2), mod)
 #' plot_zvelocity(fit)
 #' @export
 plot_zvelocity <- function(x, width = 500, height = 520, ...) {
@@ -176,12 +204,22 @@ plot_zvelocity <- function(x, width = 500, height = 520, ...) {
   xlab <- hbgd::hbgd_labels[[x$x_var]]
   ylab <- paste(hbgd::hbgd_labels[[x$y_var]], "z-score growth velocity")
 
-  figure(width = width, height = height, xlab = xlab, ylab = ylab, ...) %>%
-    ly_lines(x, dz, data = x$fitgrid, color = "black")
+  # remove blip in velocity
+  xx <- x$fitgrid$x
+  dzz <- x$fitgrid$dz
+  ind <- which.min(abs(xx - 365.25 * 2))
+  if(abs(365.25 * 2 - xx[ind]) < 2 * diff(xx[1:2])) {
+    dzz[max(1, ind - 2):min(length(xx), ind + 2)] <- NA
+  }
+
+  figure(width = width, height = height,
+    xlab = xlab, ylab = ylab, logo = NULL, ...) %>%
+    ly_lines(xx, dzz, color = "black")
 }
 
 empty_plot <- function(lab) {
-  figure(xaxes = FALSE, yaxes = FALSE, xgrid = FALSE, ygrid = FALSE) %>%
+  figure(xaxes = FALSE, yaxes = FALSE,
+    xgrid = FALSE, ygrid = FALSE, logo = NULL) %>%
     ly_text(0, 0, c("", lab), align = "center")
 }